[FEAT] AI Service Controller 분리
[FEAT] AI Service Controller 분리
