+++ WhyNComing/.gitignore
@@ -42,3 +42,4 @@ out/
 application-secret.yml
 application-jwt.yml
 application-toss.yml
+application-ai.yml
